requirements.txt
fastapi
uvicorn
boto3
pandas
pyarrow
python-dotenv
scikit-learn
tqdm
awscli
gunicorn
pytest
mangum